refactor(testing): iter 5 - deduplicate whitelist logic
the whitelist resolution in waitUntilReactiveTasksIdle and
assertNoReactiveTaskErrors was a 40-line copy-paste with subtly different
handling of edge cases (empty rule list, every-rule-no-matches). both
utilities now delegate to a single helper: resolveWhitelistFilter. the
helper returns a three-state result so callers can express 'skip',
'no filter' and 'filter' outcomes without re-implementing the logic.

the shared WhitelistRule type is exported from 'mongodash/testing' so
consumers can build reusable whitelist definitions. this also removes
the eslint-disable for `as any[]` in both files.

behavior preserved: same test cases, same results.
