+++ src/testing/index.ts
@@ -1,4 +1,5 @@
 export * from './assertNoReactiveTaskErrors';
 export * from './configureForTesting';
+export * from './resolveWhitelistFilter';
 export * from './waitUntil';
 export * from './waitUntilReactiveTasksIdle';
